+++ src/app/api/imageUrlToBase64/route.ts
@@ -1,5 +1,5 @@
 import { MOCK_BASE64 } from '@/mock/mockBlob'
-import { ImageUrlToBlobResponse } from '@/server/imageUrlToBlob/fetchImageUrlToBlob'
+import { ImageUrlToBase64Response } from '@/server/imageUrlToBase64/fetchImageUrlToBase64'
 import { mockedEndpoints } from '@/server/mockedEndpoints'
 import { ApiResponse } from '@/server/response'
 import { convertImageToBase64 } from '@/utils/convertImageToBase64'
@@ -8,10 +8,10 @@ import { NextRequest, NextResponse } from 'next/server'
 export const GET = async (
   req: NextRequest,
   res: NextResponse,
-): ApiResponse<ImageUrlToBlobResponse> => {
+): ApiResponse<ImageUrlToBase64Response> => {
   if (
     process.env.USE_MOCK_ENDPOINTS === 'true' ||
-    mockedEndpoints.imageUrlToBlob.get
+    mockedEndpoints.imageUrlToBase64.get
   ) {
     return NextResponse.json(
       {
