+++ src/server/imageUrlToBase64/fetchImageUrlToBase64.ts
@@ -1,0 +1,14 @@
+import { fetchFromApi } from '../response'
+
+export type ImageUrlToBase64Response = {
+  base64: string
+}
+
+export type ImageUrlToBase64Request = {
+  url: string
+}
+
+export const fetchImageUrlToBase64 = fetchFromApi<
+  ImageUrlToBase64Request,
+  ImageUrlToBase64Response
+>('imageUrlToBase64', 'GET')
