--- src/server/imageUrlToBlob/fetchImageUrlToBlob.ts
@@ -1,14 +1,0 @@
-import { fetchFromApi } from '../response'
-
-export type ImageUrlToBlobResponse = {
-  base64: string
-}
-
-export type ImageUrlToBlobRequest = {
-  url: string
-}
-
-export const fetchImageUrlToBlob = fetchFromApi<
-  ImageUrlToBlobRequest,
-  ImageUrlToBlobResponse
->('imageUrlToBlob', 'GET')
