+++ src/server/routes.ts
@@ -1,2 +1,2 @@
-export const API_ROUTES = ['generateImage', 'imageUrlToBlob'] as const
+export const API_ROUTES = ['generateImage', 'imageUrlToBase64'] as const
 export type ApiRouteUrl = (typeof API_ROUTES)[number]
